+++ example/presentation/lib/main.dart
@@ -1,6 +1,10 @@
+import 'package:di/di.dart';
 import 'package:flutter/material.dart';
+import 'package:get_it/get_it.dart';
 
-void main() {
+Future<void> main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await initDi(get: GetIt.instance);
   runApp(const CleanArchitectApp());
 }
 
@@ -15,9 +19,7 @@ class CleanArchitectApp extends StatelessWidget {
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.blue),
         useMaterial3: true,
       ),
-      home: const Scaffold(
-        body: Center(child: Text('Clean Architect')),
-      ),
+      home: const Scaffold(body: Center(child: Text('Clean Architect'))),
     );
   }
 }